% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{iliamna_totalcounts}
\alias{iliamna_totalcounts}
\title{Harbor Seal Aerial Survey Counts, Iliamna Lake}
\format{\describe{
\item{datetime}{survey date (and, when available, time) in Alaska local time}
\item{agency}{agency with primary responsibility for conducting the survey}
\item{observer}{individual observer(s) with primary responsibility for conducting the survey}
\item{adulttotal}{integer value representing the total number of adult seals counted across all sites}
\item{puptotal}{integer value representing the total number of pups counted across all sites}
\item{airtemp}{categorical value representing observed temperature range (F) during flight}
\item{windspeed}{categorical value representing observed wind speed (mph) range during flight}
\item{winddirection}{categorical value representing observed, prominent wind direction (cardinal) during flight}
\item{skycondition}{categorical value representing observed sky conditions during flight}
\item{precip}{categorical value representing observed precipitation activity during flight}
\item{notes}{caveats and other notes recorded during the survey or during data collation}
}}
\usage{
iliamna_totalcounts
}
\description{
Harbor Seal Aerial Survey Counts, Iliamna Lake
}
\keyword{datasets}

