% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{iliamna_waypoints}
\alias{iliamna_waypoints}
\title{Lake Iliamna Harbor Seal Haul-out Locations}
\format{\describe{
\item{wptid}{A unique character identifier}
\item{latitude,longitude}{Location of the waypoint; geographic coordinates, WGS84}
}}
\source{
\url{http://data.noaa.gov}
}
\usage{
iliamna_waypoints
}
\description{
Assembled and collated by NMML.
}
\keyword{datasets}

