% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{iliamna_waypoints}
\alias{iliamna_waypoints}
\title{=Harbor Seal Haul-out Locations, Iliamna Lake}
\format{\describe{
\item{wptid}{A unique character identifier}
\item{latitude,longitude}{Location of the waypoint; geographic coordinates, WGS84}
}}
\usage{
iliamna_waypoints
}
\description{
=Harbor Seal Haul-out Locations, Iliamna Lake
}
\keyword{datasets}

